Split out circle style plugin into generic junit reports plugin
