--- gradle-circle-style/src/main/java/com/palantir/gradle/circlestyle/FailuresSupplier.java
@@ -1,26 +1,0 @@
-/*
- * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.palantir.gradle.circlestyle;
-
-import java.io.IOException;
-import java.nio.file.Path;
-import java.util.List;
-
-interface FailuresSupplier {
-    List<Failure> getFailures() throws IOException;
-
-    RuntimeException handleInternalFailure(Path reportDir, RuntimeException ex);
-}
