--- gradle-circle-style/src/main/java/com/palantir/gradle/circlestyle/ReportHandler.java
@@ -1,29 +1,0 @@
-/*
- * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.palantir.gradle.circlestyle;
-
-import java.util.List;
-import org.gradle.api.Task;
-import org.gradle.api.reporting.ReportContainer;
-import org.gradle.api.reporting.Reporting;
-import org.gradle.api.reporting.SingleFileReport;
-import org.xml.sax.helpers.DefaultHandler;
-
-abstract class ReportHandler<T extends Task & Reporting<? extends ReportContainer<SingleFileReport>>>
-        extends DefaultHandler {
-    public abstract void configureTask(T task);
-    public abstract List<Failure> failures();
-}
