+++ gradle-junit-reports/src/main/resources/META-INF/gradle-plugins/com.palantir.junit-reports.properties
@@ -1,0 +1,1 @@
+implementation-class=com.palantir.baseline.plugins.BaselineCircleCi
