+++ gradle-junit-reports/src/test/resources/com/palantir/gradle/junit/empty-checkstyle-report.xml
