+++ gradle-junit-reports/src/test/resources/com/palantir/gradle/junit/no-failures-checkstyle.xml
