+++ gradle-junit-reports/src/test/resources/com/palantir/gradle/junit/two-exit-errors-findbugs.xml
