+++ gradle-junit-reports/src/test/resources/com/palantir/gradle/junit/two-namecheck-failures-checkstyle-report.xml
