+++ settings.gradle
@@ -4,4 +4,4 @@ enableFeaturePreview("STABLE_PUBLISHING")
 include "baseline-error-prone"
 include "gradle-baseline-java"
 include "gradle-baseline-java-config"
-include "gradle-circle-style"
+include "gradle-junit-reports"
